web/requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework==3.8.2
django-rest-auth==0.9.3
oauthlib>=3.0.0
django-allauth==0.37.1
djangorestframework-simplejwt==1.3
